Refactor `Location` to use `Position`
Introduced a `Position` data class to encapsulate line/column details
and updated `Location` to use `Position` for its `start` and `end`
fields. This should make it harder to run into small mistakes as in
commit d2ccb9a7.
The factory method `Location.create` is added to easily refactor the
existing code and keep the possibility open to create Locations without
using the `Position` data class.