Create a VSCode Desktop Extension
With the language server created in #125, this commit introduces a
very simple vscode extension for the desktop.
For now, it can only be started by launching the `Run Extension` in
`launch.json`.
There is some funky behavior for adjusting the formatting options. It
there are two formatting options to set (indentSize and insertSpaces).
The `indentSize` does seem to come through to the `FormattingService`
after a change in `./.vscode/settings.json`. However, `insertSpaces`
always stays false.