Add TM Grammar and improve language-configuration
For tokenization we are using the `SemanticTokenService` of the LSP.
The benefit of that is that we get the great tokenization of our Kson
parser.
However, when configuring the plugin to only auto-complete outside of
comments and strings it didn't work. Turns out that the vscode
`language-configuration.json`needs a TextMate grammar to identify
strings and comments.