Try building a vsix plugin
I naively tried building a plugin, by running `vsce package`. However,
that does not get the desired result. It seemed to work with the test
runner, however, when loading the plugin in an external vscode it only
executed the tasks defined in the language-configuration. Meaning that
the LSP was not started.
I'm not sure why the tests did not fail, my guess is that the reason is
that the tests were run from the project root and vscode was able to
find a correct entrypoint. Alternative explanation is that it didn't
even use the plugin, but the extensionDevelopmentPath instead.