Bundle the plugin
My naive approach in the previous commit did not work. I stumbled upon
the following guide: https://code.visualstudio.com/api/working-with-extensions/bundling-extension.
I followed this to create a bundled vscode plugin. Since we bundle
everything, and specify the entrypoints for the client and server it now
not only works with the test, but also while installing it in a regular
vscode desktop!