Fully replace old testing strategy
There were two testing strategies mixed. Initially I was using
test-cli. However, that doesn't allow us to install our own extension
before running the test. Creating your own test runner, gives more
flexibility. In this commit we remove the old approach.
The test are run with an environment variable to specify whether we are
running/testing the with or without the vsix plugin.

Note: maybe we are not even interested in running without the plugin,
then some of the complexity in this commit can be removed