Configure plugin as gradle subproject
This project is now a gradle subproject and is run and tested with
`./gradlew check`.

Note: the e2e tests require a download of vscode and are quite
slow, since we use timeouts for tasks to happen. Even if they could be
faster it is probably undesirable to spin-up a vscode for testing if you
are working on a feature unrelated to this plugin.