Refactor BackspaceHandler and TypedHandler
I realized that `KsonBackspaceHandlerDelegate` overrides the
`beforeCharTyped` method while the `KsonTypedHandlerDelegate` overrides
the `charTyped` method. It turns out that using `beforeCharTyped` is
easier to implement, especially when we need to check whether the caret
is in a specific element.
I remembered how confusing it was to work with the caret offset when
fixing a bug where we auto-completed in a comment (see: aa7bdb35).
Overriding the `beforeCharTyped` makes this a lot more intuitive.

Another place where we used the `caretOffsets` was to find the character
that would be deleted or typed. This character is passed as an argument,
so we don't need the `text[caretOffset]` or `text[caretOffset - 1]`
anymore. Instead of the `if` statements we use a `when` statement
matching the typed or deleted character.