Refactor build and tests for browser
To prepare ourselves for adding browser compatibility to the tests and
plugin we need to refactor in such a way that we can extend our current
set-up to support both runtimes.

The build file, `esbuild.js`, is split in such a way that we can add
entrypoints depending on the runtime. The `index.ts` test index is
refactored to `index.common` so that we could easily extend and
customize the imports per runtime in `index.node` and `index.browser`.