Add tests for browser functionality
Note: Currently the browser tests are failing we still need to add
browser support for VsCode. This will be done in future commits.
The tests are executed by starting vscode on the browser, loading up the
extension and executing.
The esbuild adds browser specific endpoints, and the files in `test/suite`
are adapted to properly handle both node and browser environments.