Add browser support for vscode plugin
The tests added for browser support are now passing! Meaning that we
are also able to host a plugin in the browser now.
To allow for browser support the server needs to be loaded up in a web
worker. In `browser/ksonServerMain` a server is created with a connection
that can be used in a webworker. In `browser/ksonClientMain` the server
is loaded in a `Worker` and with that worker a language client is
constructed.