Test to ensure optional comma precedence
Add a test to ensure our precedence for optional commas: a comma between
two object properties in a bracket-list belongs to the object, not the
list.  i.e. this value:

```
[
  k1: v,
  k2: v
]
```

is always parsed as a list of one object with two properties:

```
- k1: v
  k2: v
```

The grammar for objects already specifies this fairly clearly, but it's
nice to have a test which explicitly ensures this remains true in the
face of future changes since it is a subtle and important distinction.