Helpful errors for reserved word object keys
Improve our parsing and error messaging when null/true/false are used
as keywords.  We used to stop parsing early in this case because
we'd parse the reserved word then bump into an unexpected colon.
We now allow parsing to continue as if this were a legal keyword,
marking it with a helpful and precise error.