Formatting for docs with illegal trailing content
If a Kson document is fairly well-formed but has some unexpected
trialing content, we should still be able to format the valid portion of
the document.  This also offsets the illegal trailing content, likely
helping a user see and understand the error more easily.

Contrived example: given a syntactically noisy case like

```
{key:[{}]}[]
```

we are now able to format, yielding clarity for both legal value and the
illegal trailing `[]`:

```
key:
  - {}

[] # <-- this is marked as "Unexpected trailing content"
```