Add Monaco LSP Client
After adding support for the browser in 8fb3bb1c, we know add support
for Monaco. My initial idea was to simply use the created `.vsix` file
and load that up in Monaco.
Theoretically that should work, however, it turned out to be incredibly
difficult to figure out why it didn't. Instead we are now using an
approach where the language client and server are wrapped.
To get it working there is some code-duplication with the approach for
vscode. This can and will be refactored in next commits.