Refactor Schema parsing Kson interface
Refactor our schema handling to parse schemas separately from general
Kson.  This is more natural as it allows Schema parse errors to be
surfaced much more naturally, and it allows clients to parse a schema
once and apply it to many Kson documents.

This also cleans up the Schema parsing code somewhat as its new home is
a more natural fit and naturally avoids a circular dependency we were
slightly awkwardly working previously.

Not that we also needed to rename `valid` to `isValid` in the
`JsonSchema` interface, as once it was `@JsExport`ed, we got a warning
about a name clash around `valid` (likely from some JS built-in?)