Clearly define Kson's public interface
Introduced the `lib-kotlin` module to clearly define and control Kson's
public API surface.  This is the authoritative owner of what our public
api contract is around the Kson functionality defined in this project.

The `Kson` class defined in `lib-kotlin` is designed specifically to be
friendly and usable in Kotlin Multiplatform, and we will likely use this
public module as the basis for a `lib-native` and possibly a
`lib-javascript` module to facilitate exposing this interface in
structures feel idiomatic and friendly on those platforms

NOTE: this disables the `tooling:language-server-protocol` and
`tooling:vs-code` modules by commenting them out in `settings.gradle.kts`
The next commit will re-nable those and refactored them to use this new
public interface