Evaluate `stringContent` in `KsonApi`
To account for unprocessed characters in string inputs, the `KsonApi`
now returns the `processedStringContent` instead of the `stringContent`.
This change is aimed at resolving issues caused by unprocessed escape
sequences, Unicode characters, and control characters in the
string content.
Specifically we ran into an issue with this when evaluating a
patternProperty from a JsonSchema that contained '\/' in the JsonSchema.
Passing that unprocessed to Regex results in a syntax error. First
processing or evaluating the string solves this class of issues.