Split-up `check` in build to solve OOM-errors
Running `./gradlew check` locally didn't result in any issues. However,
running it in CircleCi did. We failed with the error message: ```
* What went wrong:
Gradle build daemon disappeared unexpectedly (it may have been killed or
may have crashed)
```
The reason for that seems to be Out-Of-Memory errors. Splitting up the
tasks and running `lsp-clients` without daemon seems to solve the issue.