Improve public `TokenType` interface
Create a public enum for the TokenTypes we want to return in `analyze`.
These are mostly the same as the underlying internal token types, but
we do simplify some string tokens that are only refined for internal
uses, and it's important to be free to make other internal refinements
in the future without needing to change this public enum.