Improve loading test files
The pattern we used before to load test files to be both browser and
server compatible was error-prone.
A new test file needed to be added in `index.common.ts` and `esbuild`.
The new approach improves that by adding a `getTestEntries()` function
during the build which generates a `test-files.json`. The
`test-files.json` contains the filenames of the tests that we can then
load in the `browser` and `node` specific way in `index.common.ts`.