Add different KSON `FormattingStyle`'s
Instead of always formatting Kson with as little delimiters as possible
we now add an option to format it with delimiters.
In this mode objects are formatted with delimiting curly braces and lists
with angle braces. This gets close to the way we would format in case of
using Json as a `CompileTarget`. The differences are:
- Strings stay unquoted or in single quotes where possible
- Lists are formatted with angle brackets and dashes
- Comma's are never added
- Embed blocks are written out as is

The benefit of this new `FormattingStyle` is that it gives a clearer and
delimited 'lens' to edit data, making it easy for instance to move
between these two views of the data:
```
list:
  - 1
  - 2
  - key: value
  -
     - 'new list'
     =
  - test
embed: $tag
  content
  $$
```
```
{
  list: <
    - 1
    - 2
    - {
        key: value
      }
    - <
        - 'new list'
      >
    - test
  >
  embed: $tag
    content
    $$
}
```