Don't run Monaco Plugin Build
The 'check' task depended on `npm_run_buildPlugins` which builds both
the Monaco and VS Code plugin.
Monaco is build with Vite, which on CircleCI, runs very slowly and
resulted in OOM-errors. In 8104127d I solved this by splitting up the
circleCI tasks.
However, the long build time is quite annoying still, since
nothing depends on Monaco building we take this out off the build and
revert the changes in 8104127d.