Add SimpleListIterator to `lib-kotlin`
This class is necessary to allow FFI users to iterate through lists
returned by kotlin. It is not _really_ part of the public API, but there
is no better place for it right now. Maybe we could introduce a
`lib-native` that exposes everything from `lib-kotlin` and adds the
`SimpleListIterator` type on top, to keep things cleaner.