Unify quoted string and unquoted string in grammar
Both quoted string and unquoted string are now seen as a string in the
grammar. This also fixes a bug where the formatter would format:
```
key : value
```
to
```
key

:

value
```.

Because in the `Parser` we looked ahead only 1 token to find a colon,
in the previous example this would have been whitespace.