Add errors for deceptively indented objects/lists
With our auto-formatter, we already do a very good job of _encouraging_
the structure of Kson files to match the structure of the actual data,
but unlike Yaml with its significant whitespace, we were not helping
_ensure_ the data's structure was well-represented.  Until now.

We now validate that the visual indentation of a Kson file does not
misrepresent the structural nesting of its data, giving an error when
object properties or list elements are indented inconsistently.

Note: this is not yet surfaced in our IntelliJ plugin (coming soon), but
does flow through to our VS Code support.