Language Injection in LSP clients
Instead of handling `EMBED_CONTENT` in the `SemanticTokensService` we
updated the textmate grammar to handle this. This way we get 'free'
language injections depending on the installed plugins in your VS Code
or Monaco.

In Monaco we need to manually add services for the supported languages,
which increases the buildsize even more. In VS Code, many of the
languages seem to be present already.

To simplify the generation of the pattern I've added a generation script
that can update the embed block part of the textmate grammar.

To sanity check that we tag and include the `sources.extension` file for
a language I've added a simple `syntax-highlighting.test` for TextMate
grammars.