Make Monaco Editor configurable
Working on the KSON website I realized that it wasn't possible to
configure Monaco editor easily. So I removed the `wrapperKson` that just
started the editor, the `KsonEditorManager` takes a config argument and
 from there we get back a `MonacoLanguageClientWrapper` object that can
 be started in for example the website.

The `index.html` and corresponding playwright test have been simplified
and updated to use this new API.

I also realized we don't need a `main` in the package.json.