Prohibit commas in plainObjects
It seemed intuitive to apply the same rules for plain objects as the
internals of delimited objects, but in practice this created unneeded
complexity and ambiguity.  In https://github.com/kson-org/kson/pull/137,
we tried to disambiguate who "owns" a comma in a plainObject inside a
bracketList by embracing plainObjects having higher precedence in the
existing code.  However, plain KSON is very similar to YAML, and YAML
would give the comma to the bracket list in this case.

With this change, KSON and YAML both agree that, for instance:

```
[
  one: 1,
  two: 2
]
```

parses to a list of two objects of size one:

```
- one: 1
- two: 2
```

Additionally, plain objects are now symmetric with dash lists, just with
a `keyword` denoting a ksonValue rather than a `LIST_DASH`, making the
core/default KSON syntax more consistent, as we can see from these two
KSON grammar rules:

```
* dashList    -> (LIST_DASH ksonValue)+ "="?
* plainObject -> (keyword   ksonValue)+ "."?
```