Bugfix: return early when creating AST fails
When a KSON file has an error the ast would be null. This means we
cannot do any validations since there is no KsonValue to validate
against.

The validation for `JsonSchema` did not have the conditional for
`ksonValue != null`, so we ran into NPE's. Returning early prevents
these errors and removes the need to cast the ast to KsonValue.