Distinguish Post-processing and Parsing Messages
`Message`'s can be created during Parsing or post-processing.
Post-processing messages are created by any of the validators, for
example `org.kson.validation.IndentValidator` or
`org.kson.schema.JsonSchemaValidator`. If the messages are created
during parsing `coreParseMessage` is set to `True`.

This distinction comes in useful in the `KsonValidatorAnnotor` where we,
only want to annotate messages that have not been annotated yet during
parsing, to prevent 'duplicate' errors.