default Message.coreParseMessage to false
Instead of defaulting Message.coreParseMessage to true and setting it to
false through the MessageType.create() method we use an extension
function in the KsonBuilder.
The advantage of this approach is that callers of MessageType.create()
never have to think about whether, a message is a coreParseMessage.
If and only if it passes through KsonMarker.error() the extension
function, `Message.asCoreParseMessage` is called and the message becomes
a coreParseMessage.