Distinguish between Warning and Errors
Instead of marking every message as an Error this commit allows to mark
some messages as Warnings.
This commit sets up the infrastructure to add different severities to
messages, by default every message is an error, and we should go through
each message case-by-case to decide to which class it belongs.