Add warnings and errors to IJ plugin
With the work done in the previous commit we are able to add different
highlighting to Messages depending on the Message severity. The tricky
part here is that every 'error' found during parsing (with
`KsonParser.error`) is marked as an error in IJ. Hence, we add a
conditional for the `Message.type.severity`. The 'un-annotated' messages
are later marked in `KsonValidationAnnotator.apply`