Add build script to use GraalVM for native builds
Because Clikt isn't fully MPP (it only supports Path and File on the
JVM), we decided to create a native binary that can be run as CLI.

Ideally we'd use the graal plugin and let the gradle toolchain plugin
handle managing the correct version of Graal. However, it turned out to
be near impossible to point the toolchain-plugin to the folder that
stores jdk's. Which is necessary since it's important for us that the
JDK is contained within the project.
Setting GRAAL_HOME as environment variable complained about not finding
a java executable (which according to stackoverflow shows up when not
configuring JAVA_HOME).
The route I took in the end is downloading graal directly, and using a
cli command to build the native image.

So even though it's more complex than I hoped (except when we make the
whole project rely on Graal we are able to natively build the CLI!