Remove ignored $ref sibling handling error
Hopefully we can add some form of this back later, but for now this is
not quite correct since these siblings are not totally ignored and may
be used for instance in some fancy $ref lookups.  Hopefully we can
revisit adding some helpful messaging in here once full $ref support
is implemented.