buildSrc: simplify `testEnsureCleanGitCheckoutOnEmptyDir`
File deletion causes error on Windows, probably due to the way JGit
implements the git clone operation. We spent more than an hour hunting
the bug down, but didn't achieve a solution. Instead of sinking more
time into it decided to simplify the test in the end so it doesn't need
to delete files at all (while still testing the same functionality).