Solve .copy() visibility warnings
Instead of exposing the non-public primary constructor we added `@ConsistentCopyVisibility`, which now ensures the visibility of copy is the same as that of the constructor.

```
This will become an error in Kotlin 2.2.
w: file:/org/kson/Kson.kt:265:20 Non-public primary constructor is exposed via the generated 'copy()' method of the 'data' class.

The generated 'copy()' will change its visibility in future releases.

To suppress the warning do one of the following:
- Annotate the data class with the '@ConsistentCopyVisibility' annotation.
- Use the '-Xconsistent-data-class-copy-visibility' compiler flag.
- Annotate the data class with the '@ExposedCopyVisibility' annotation
  (Discouraged, but can be used to keep binary compatibility).
 ```