Enhance AnyOf Schema error messaging
Rather than an opaque "must match at least one of the specified schemas"
on the whole offending object, we now bubble up specific errors,
anchoring them to specific KSON source when possible (i.e. when an
error is applicable for all anyOf sub-schemas)

This may yet be further improved, but this is a good step forward that
also makes a good example of what's possible with regards to schema
errors in KSON as we progressively polish the messaging across the
entire schema validation system

This change also involved enhancements our Message and KsonObject
interfaces to facilate this type of better error messaging. Messages
are now trivial to de-dupe by using a Set, and KsonObject values now
make it much easier to report error by location of their keys.