Allow builds from source distribution
This is a first step to creating wheels for platforms. The first
priority is that each platform can built the kson library from source.
In the `build.gradle.kts` we add a task that copies the necessary files
to a directory (instead of just cloning everything). The `MANIFEST.in`
specifies which files to include in the dist (maybe this is double since
we are detailed about what to include in kson-sdist.
Copying and creating the dist zip file is possible with
`./gradlew :lib-python:createSdistBuildEnvironment`.

The custom `build_backend.py`  Will move into that dist, run the gradlew
task necessary to create the bindings and copies the bindings and the
artifacts into the kson library.

The build directory becomes pretty large so this is removed as last step
in the build-backend.py