Rename lib-kotlin to kson-lib
Calling this module `kson-lib` better highlights its importance as the
public interface of the kson project. This module is the key source
of truth for what is exposed by this project and owns for instance the
"source" version the interface's documentation.

Furthermore, hopefully with this new naming `lib-python` and `lib-rust`
more immediately register as specializations of this lib depending on
`lib-kson`, rather than pure siblings to it.