Index and cache the `DocumentSymbols`
The `IndexedDocumentSymbol` prepares for adding a `HighlightingService`.
In the `HighlightingService` we need to retrieve all symbols at a certain position and select the position that is the 'smallest'.

The `collectAndIndexSymbols` creates a mapping of line numbers to all
`DocumentSymbols` on that line. This way we can do a quick lookup for
the highlighting.

To not redo these computations often we 'cache' these results in the `KsonDocument`.