Use cibuildwheel to build wheels
Using cibuildwheel we generate the wheels for different platforms. It
seems like we need a few files as `setup.cfg` to set the python api we
are building from, and `_marker.c` to 'trick' it into seeing that this
is not a pure python package (in which case it would build a wheel for
any platform).

cibuildwheel also immediately tests the wheel on the smoketest.

Add a task that builds wheels

After a successful installation from sdist we build the wheel. This
should be pretty quick since the artifacts are present already.