Add twine to push to pypi
Using `twine` we can push a distribution to pypi manually. Using the
following command:
`./uvw run python -m twine upload --repository testpypi dist/* --verbose`.
It requires us to put the wheels and the sdist in the folder to publish.
Wheels are currently built at CI, downloaded and copied into the folder
to publish.