safeguard against accidental overwrites
To prevent accidental overwrites of your YAML
file I've added some safeguards to the transpile
command. Before it could have happened that changes
are made to the `YAML` file, which are overwritten
by running any gradle command. If there is a git
repository present we are 'safe' from these mistakes
now by checking for uncommitted changes.
On top of that we only are able to run the task if
the files are committed in-sync.
The cost is that you can't continuously 'run' the
transpile kson to yaml task, without 'undoing' the
changes to yml or committing.
I think that's a price that I'd happily pay for
preventing accidental overwrites.