Use a value-based "logical" equals for KsonValue
We toyed with being more strict in d05967d3, but in practice we had it
right initially: KsonValue equals should be implemented as a logical
equals based on the underlying value.

This makes a KsonValue tree much more ergonomic to work with because
the values really then become _values_ that can be compared as such.