Fix formatter bug with trailing content
This content
```kson
outer1:
  nest1:
    nest2: 4
    .
  .
. # extra end-dot
outer2: 8 # unexpected trailing content
```
Got formatted to:
```kson
# extra end-dot
outer1:
  nest1:
    nest2: 4

outer2: 8 # unexpected trailing content
```
Meaning that the data of the document changes meaning. The correct way
to format would be to keep the end-dots. A similar problem occured with
end-dashes.
We solved this by passing `trailingContent` as `nextNode` in `KsonRoot`.
If the nextNode is a ksonValue we know that there is trailing content
and we preserve the end-dashes and end-dots.