Build document path
Build a document path given a location and the document as string. If
the document is invalid we try to 'recover' by creating a string node at
the location.

The 'hard' part is where to navigate to. Let's say we are in the
following position:
```kson
key: <caret>
```
Then we should navigate to 'path'. However, if we are typing 'key':
```kson
ke<caret>
```
or finished with the key and value:
```kson
key: "value"<caret>
```
Then we should navigate to the root of the document.
This is solved by using the tokens of the document to find the correct
path.