Add schema loading through filesystem
Add the possibility to configure schema's in VS Code through a config
file. This is a good step to getting feature parity with plugins like
YAML Redhat, where many of the benefits are dependent on the schema
being configured.

As an MVP I've written a `FileSystemProvider` that can be configured
with a `.kson-schema.kson` file. In this file you can map kson-files
to a schema.
The schema is mapped to a `KsonDocument` through the schema provider in
`KsonDocumentsManager`.
One 'problem' is that we are not able to use this in the browser
currently because we are using the filesystem. My idea is that by adding
different Providers (SchemaStoreProvider for example) we could also get
similar support in the browser.
To make sure the `FileSystemProvider` is only used in cases when on
node I've created a different entrypoint for both browser and
node through `./index.node.ts` and `./index.browser.ts`, which can be
used by the desktop and browser version of the VS Code plugin.