Associate schema's through status bar in VSCode
This commit allows us to add schema's to a file through the status bar.
If a schema has already been connected through the schema-config file,
the status bar shows this as well.

Although not implemented yet, it should be possible to start 'removing'
mapped schema's.

Another todo would be to allow for URL-schema's to be mapped with a file.

If the schema-config file doesn't exist yet it gets generated.

On the LSP/server-side we added an `AssociateSchemaCommand` and a
`getDocumentSchema` endpoint. The `getDocumentSchema` endpoint is called
to show the correct schema in the status bar.
The `AssociateSchemaCommand` is used to add or remove a schema to a file.