Change `KsonObject` in public API
This change makes it easier to retrieve the `propertyKeys` and
`propertyValues` from a KsonObject. Before the way to retrieve
`KsonValues` in an object was to make a mapping before being able to
look-up based on key.

Now this is part of the api. To retrieve the keys with locations or
values one can simply use.  `object.properties.get('key')` or `object.
propertyKeys.get('key')`.