Always use the latest kson-core JAR in graalvm native image build
Instead of hardcoding the path, which changes every time we bump KSON's
version, we now retrieve it directly from the corresponding Gradle task.

Additionally, we now check that all input JAR files actually exist, and
raise an exception when any of them is missing. Otherwise, we see
puzzling errors down the road (graalvm is unfortunately too lax and
compiles our code even if some of the specified JAR files are missing,
leading to hard-to-debug failures in our test suite).