add `JsonPointerPlusParser`
`JsonPointer`doesn't allow wildcards in the pointer. That is
not an issue for `JsonPointer`'s in a schema. However, if we
want to navigate a KsonValue using more expressive pointers we
can't.

Options to deal with this would be using `JsonPath`, `JmsePath`.
However, before we commit to that I'd propose we use a simple
improvement of `JsonPointer`, namely `JsonPointerPlus`. It adds
the possibility to use wildcards `*`, and glob matching for a
property key (only with ? (matching a single character) and *
(matching all characters)).